Fix/modal editar rating mark

+++ frontend/components/modals/RateModal/UpdateRatingProduct.tsx
@@ -2,17 +2,22 @@ import { useEffect, useState } from "react";
 import RateModal from "./RateModal";
 import { toast } from "react-toastify";
 import { getProductRating, updateProductRating, deleteProductRating } from "@/api/api";
+import { on } from "events";
+import { useRouter } from "next/navigation";
 
 interface Props {
   ratingId: number;
   open: boolean;
   onClose: () => void;
+  onSuccess: () => void;
 }
 
-export function UpdateProductRatingModal({ ratingId, open, onClose }: Props) {
+export function UpdateProductRatingModal({ ratingId, open, onClose, onSuccess }: Props) {
+  const router = useRouter();
   const [rating, setRating] = useState<number>(0);
   const [comment, setComment] = useState<string>("");
   const [productName, setProductName] = useState<string>("");
+  const [productId, setProductId] = useState<number | null>(null);
 
   useEffect(() => {
     async function loadRating() {
@@ -22,6 +27,7 @@ export function UpdateProductRatingModal({ ratingId, open, onClose }: Props) {
         setRating(res.rating);
         setComment(res.comment);
         setProductName(res.product.name);
+        setProductId(res.product_id);
 
       } catch (err) {
         toast.error("Erro ao carregar avaliação.");
@@ -37,6 +43,7 @@ export function UpdateProductRatingModal({ ratingId, open, onClose }: Props) {
     try {
       await updateProductRating(ratingId, { rating, comment });
       toast.success("Avaliação atualizada!");
+      onSuccess();
       onClose();
     } catch {
       toast.error("Erro ao atualizar avaliação!");
@@ -45,10 +52,12 @@ export function UpdateProductRatingModal({ ratingId, open, onClose }: Props) {
 
   // ---- DELETAR ----
   const handleDelete = async () => {
+    if (!confirm("Deletar avaliação?")) return;
     try {
       await deleteProductRating(ratingId);
       toast.success("Avaliação removida!");
       onClose();
+      router.push(`/product/${productId}`)
     } catch {
       toast.error("Erro ao remover avaliação!");
     }
@@ -57,7 +66,7 @@ export function UpdateProductRatingModal({ ratingId, open, onClose }: Props) {
   return (
     <RateModal
       open={open}
-      title={`Editar: ${productName}`}
+      title={`${productName}`}
       rating={rating}
       setRating={setRating}
       comment={comment}
