+++ frontend/components/modals/RateModal/UpdateRatingStore.tsx
@@ -1,19 +1,24 @@
-import { useEffect, useState } from "react";
+import { use, useEffect, useState } from "react";
 import RateModal from "./RateModal";
 import { toast } from "react-toastify";
 import { getStoreRating, updateStoreRating, deleteStoreRating } from "@/api/api";
 import { on } from "events";
+import { useRouter } from "next/navigation";
 
 interface Props {
   ratingId: number;
   open: boolean;
   onClose: () => void;
+  onSuccess: () => void;
 }
 
-export default function UpdateStoreRatingModal({ ratingId, open, onClose }: Props) {
+export default function UpdateStoreRatingModal({ ratingId, open, onClose, onSuccess }: Props) {
+  const router = useRouter();
+
   const [rating, setRating] = useState<number>(0);
   const [comment, setComment] = useState<string>("");
   const [storeName, setStoreName] = useState<string>("");
+  const [storeId, setStoreId] = useState<number | null>(null);
 
   useEffect(() => {
     async function loadRating() {
@@ -23,6 +28,7 @@ export default function UpdateStoreRatingModal({ ratingId, open, onClose }: Prop
         setRating(res.rating);
         setComment(res.comment);
         setStoreName(res.store.name);
+        setStoreId(res.store_id);
 
       } catch (err) {
         toast.error("Erro ao carregar avaliação.");
@@ -38,6 +44,7 @@ export default function UpdateStoreRatingModal({ ratingId, open, onClose }: Prop
     try {
       await updateStoreRating(ratingId, { rating, comment });
       toast.success("Avaliação atualizada!");
+      onSuccess()
       onClose();
     } catch {
       toast.error("Erro ao atualizar avaliação!");
@@ -46,10 +53,12 @@ export default function UpdateStoreRatingModal({ ratingId, open, onClose }: Prop
 
   // ---- DELETAR ----
   const handleDelete = async () => {
+    if (!confirm("Deletar avaliação?")) return;
     try {
       await deleteStoreRating(ratingId);
       toast.success("Avaliação removida!");
       onClose();
+      router.push(`/store/${storeId}/ratings`)
     } catch {
       toast.error("Erro ao remover avaliação!");
     }
@@ -58,7 +67,7 @@ export default function UpdateStoreRatingModal({ ratingId, open, onClose }: Prop
   return (
     <RateModal
       open={open}
-      title={`Editar: ${storeName}`}
+      title={`${storeName}`}
       rating={rating}
       setRating={setRating}
       comment={comment}
